Mostly package/webpack/ts configuration. Slightly refactored the TS code. Package version 0.0.5
