+++ package.json
@@ -1,10 +1,12 @@
 {
-  "name": "@mindbox/quokka-ts",
-  "version": "0.0.1",
+  "name": "@mindbox/quokka",
+  "version": "0.0.5",
   "description": "Mindbox.Quokka Typescript implementation",
   "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
-    "test": "webpack && karma start"
+    "test": "karma start",
+    "prepare": "npx webpack"
   },
   "repository": {
     "type": "git",
@@ -20,9 +22,17 @@
     "url": "https://github.com/mindbox-moscow/Quokka.Ts/issues"
   },
   "homepage": "https://github.com/mindbox-moscow/Quokka.Ts#readme",
+  "files": [
+    "index.d.ts",
+    "dist"
+  ],
   "devDependencies": {
+    "@babel/core": "^7.1.6",
+    "@babel/preset-env": "^7.1.6",
     "@types/jasmine": "^3.3.0",
+    "babel-loader": "^8.0.0-beta.0",
     "karma": "^3.1.1",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-jasmine": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-typescript": "^3.0.13",
@@ -33,7 +43,7 @@
     "webpack-cli": "^3.1.2"
   },
   "dependencies": {
-    "antlr4ts": "^0.4.1-alpha.0",
-    "karma-chrome-launcher": "^2.2.0"
+    "@babel/polyfill": "^7.0.0",
+    "antlr4ts": "^0.4.1-alpha.0"
   }
 }
