+++ src/Implementation/render-context.ts
@@ -1,0 +1,4 @@
+export interface RenderContext {
+	compositeModelValue: Object;
+	functionRegistry: { [name: string]: Function };
+}
