+++ tests.tsconfig.json
@@ -1,0 +1,5 @@
+{
+    "extends": "./tsconfig",
+
+    "exclude": []
+}
