+++ tsconfig.json
@@ -1,10 +1,15 @@
 {
-    "compilerOptions": {
-      "target": "es2015",
-      "module": "commonjs",
-      "declaration": true,
-      "outDir": "./dist",
-      "sourceMap": true,
-      "experimentalDecorators": true
-    }
-  }
+  "compilerOptions": {
+    // Antlr4 runtime requires es2015
+    "target": "es2015",
+    "module": "commonjs",
+    "declaration": true,
+    "outDir": "./dist/",
+    "sourceMap": true,
+    "experimentalDecorators": true
+  },
+
+  "exclude": [
+    "**/*.spec.ts"
+  ]
+}
